requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
Django==4.2.6
django-filter==23.3
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg==3.1.12
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
whitenoise==6.6.0

